requirements.txt
# Flask及其扩展
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.0.1
Flask-Migrate==4.0.5
Flask-Bootstrap4==4.0.2
flask-apscheduler

# 数据库
SQLAlchemy==2.0.23

# HTTP请求
requests==2.31.0
requests-toolbelt==1.0.0

# 工具库
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
aliyun-python-sdk-core
aliyun-python-sdk-alidns

# 邮件发送
Flask-Mail==0.9.1

# 开发工具
pytest==7.4.3
pytest-flask==1.2.0
